Explicit typecasting

+++ data/SugarBean.php
@@ -1447,7 +1447,7 @@ class SugarBean
 
         $result = $this->db->query($count_query, true, "Error running count query for $this->object_name List: ");
         while ($row = $this->db->fetchByAssoc($result, true)) {
-            $num_rows_in_query += current($row);
+            $num_rows_in_query += (int)current($row);
         }
 
         return $num_rows_in_query;
